Extract a "conditions widget" out of the conditions-forms
